refactor(r-button-group): a dedicated folder for RButtonGroup
