+++ packages/recomponents/src/components/r-button/r-button.story.js
@@ -3,7 +3,6 @@ import {action} from '@storybook/addon-actions';
 import {boolean, select, text} from '@storybook/addon-knobs';
 import StoryRouter from 'storybook-vue-router';
 import markdown from './r-button.md';
-import markdownGroup from './r-button-group.md';
 
 storiesOf('Components/Button', module)
     .addDecorator(StoryRouter())
@@ -60,22 +59,4 @@ storiesOf('Components/Button', module)
             </r-button>`,
     }), {
         notes: {markdown},
-    })
-    .add('Button group', () => ({
-        props: {
-            active: {
-                default: boolean('active', false),
-            },
-            fluid: {
-                default: boolean('fluid', false),
-            },
-        },
-        template: `<r-button-group :fluid="fluid">
-                        <r-button>One</r-button>
-                        <r-button>Two</r-button>
-                        <r-button :active="active">Three</r-button>
-                        <r-button>Four</r-button>
-                    </r-button-group>`,
-    }), {
-        notes: {markdown: markdownGroup},
     });
